Enable support for random port binding in tomcat server.
Moves the console messages to after Tomcat is started so we can get the
actual http/https ports.
Changes the TomcatKillSwitch port to be HTTP-1
instead of +1 to avoid conflicts when both http and https are random
ports.
Adds the TestPhasePrepared event to let plugins know when Tomcat is
started, so they can get the real ports.